Split test module for metavariable expressions
